+++ packages/solverpy-learn/src/solverpy_learn/setups/__init__.py
@@ -1,4 +1,4 @@
-from solverpy.setups import Setup, vampire, prover9, bitwuzla, z3, evaluation
+from solverpy.setups import Setup, Evalset, vampire, prover9, bitwuzla, z3, evaluation
 from solverpy.setups import launch as _launch_base
 
 from .solver import eprover, cvc5
@@ -7,6 +7,7 @@ from .loop import initialize, loopinit, looping, oneloop, launch
 
 __all__ = [
    "Setup",
+   "Evalset",
    # solvers
    "eprover",
    "vampire",