+++ packages/solverpy/src/solverpy/setups/__init__.py
@@ -39,11 +39,13 @@ strategy and benchmark id lists), then call
 """
 
 from .setup import Setup
+from .evalset import Evalset
 from .solver import eprover, vampire, prover9, cvc5, bitwuzla, llm2smt, z3
 from .loop import evaluation, launch
 
 __all__ = [
    "Setup",
+   "Evalset",
    "eprover",
    "vampire",
    "cvc5",