+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased changes (2026-06-14)
 
-* refactor(setup): plan Dataset split and init ordering [[details](https://github.com/cbboyan/solverpy/commit/) | [browse](https://github.com/cbboyan/solverpy/tree/)]
+* refactor(setup): extract Evalset, Runtime, Managed; move talker into Setup [[details](https://github.com/cbboyan/solverpy/commit/) | [browse](https://github.com/cbboyan/solverpy/tree/)]
+* refactor(setup): plan Dataset split and init ordering [[details](https://github.com/cbboyan/solverpy/commit/7782d6f) | [browse](https://github.com/cbboyan/solverpy/tree/7782d6f)]
 * fix(autotune): preload validation data [[details](https://github.com/cbboyan/solverpy/commit/4989b84) | [browse](https://github.com/cbboyan/solverpy/tree/4989b84)]
 * fix(autotune): isolate repeated tuning phases [[details](https://github.com/cbboyan/solverpy/commit/16eb4d3) | [browse](https://github.com/cbboyan/solverpy/tree/16eb4d3)]
 * feat(smt): add llm2smt solver [[details](https://github.com/cbboyan/solverpy/commit/c3fc5fc) | [browse](https://github.com/cbboyan/solverpy/tree/c3fc5fc)]